refactor code for dashboard and editor page
