+++ client/package.json
@@ -14,6 +14,7 @@
     "@monaco-editor/react": "^4.7.0",
     "axios": "^1.9.0",
     "jwt-decode": "^4.0.0",
+    "prop-types": "^15.8.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router-dom": "^7.6.0",
